+++ src/environments/environment.prod.ts
@@ -1,17 +1,1 @@
-export const environment = {
-  production: true,
-  supabase: {
-    url: 'https://oktbesfgiyhcjpaljhby.supabase.co',
-    anonKey:
-      'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJzdXBhYmFzZSIsInJlZiI6Im9rdGJlc2ZnaXloY2pwYWxqaGJ5Iiwicm9sZSI6ImFub24iLCJpYXQiOjE3NjI4NzAxNDcsImV4cCI6MjA3ODQ0NjE0N30.Hqm4UoWLqwKYO49t5tu5TSft2_oOyvH8DO1Q4wzSKmA',
-  },
-  google: {
-    clientId: '1054129590650-c4qqjfphqdejcfdcasel8q3a14b5p7kk.apps.googleusercontent.com',
-  },
-  microsoft: {
-    clientId: 'YOUR_MICROSOFT_CLIENT_ID',
-    redirectUri: 'https://devpad.therama.dev',
-  },
-  apiUrl: 'https://api.therama.dev',
-  appVersion: '2.0.0',
-};
+export const environment = {};